requirements.txt
# Phase 1b runtime dependencies (pinned for reproducibility)
fast-flights==2.2
openpyxl==3.1.5
python-dotenv==1.2.2
requests==2.33.1
resend==2.29.0
sqlalchemy==2.0.49
xlrd==2.0.2
